added folder delete function
